+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased:
 
+- Add command line interface
 - Watch assets dir for changes in watch mode
 - Added `refresh-templates` task to update to latest templates
 - Social sharing (preview for Facebook, Twitter, LinkedIn, etc.)